Initialize ata before graphics
ATA init is the long pole in the boot process, and its asynchronous.
move the graphics init after it so that ata and graphics initialize
in parallel

UFS is also a part of ATA and is also somewhat slow in probing.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
(cherry picked from commit 29cc8482c994188d774af7dc5f6ba7723a81e189)