GKI: io-pgtable: ignore debugfs failure
Recognize that debugfs is not critical, so being able to set it up
should be considered optional and should not result in an init error.
Remove unnecessary checking.

The patch source of the init and exit routines fails to tag the
functions __init, __initconst or __exit respectively, fix.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 141888626
Change-Id: Ic164c172bab70b8626898d69cc906d99693e1ae3
Signed-off-by: dreamisbaka <jolinux.g@gmail.com>
(cherry picked from commit a72542872a3f1830608e1a0289bdbda54dbe1e06)