msm: kgsl: Affine IRQ and worker kthread to the big CPU cluster
These are in the critical path for rendering frames to the display, so
mark them as performance-critical and affine them to the big CPU
cluster. They aren't placed onto the prime cluster because the
single-CPU prime cluster will be used to run the DRM IRQ and kthreads.
DRM is more latency-critical than KGSL and we need to have DRM and KGSL
running on separate CPUs for the best performance, so KGSL gets the big
cluster.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit a3fc210d73e28574b2be4e86391be5e30df96407)